Move the tests to a common location to be run against multiple backends via a parametrized fixture
