+++ pyproject.toml
@@ -51,6 +51,8 @@ mike = { version="^1.1.2", optional=true}
 pyyaml = "^6.0.1"
 pyarrow = {version = "^13.0.0", optional = true}
 
+polars = {version = "^0.19.12", optional = true}
+
 [tool.poetry.extras]
 test = [
     "pytest",
