+++ tests/rules/io/test_files_csv.py
@@ -1,11 +1,10 @@
 import datetime
 import os
 from pandas import DataFrame
-from pandas.testing import assert_frame_equal
 import pytest
 
 from etlrules.backends.pandas import ReadCSVFileRule, WriteCSVFileRule
-from tests.backends.pandas.utils.data import get_test_data
+from tests.utils.data import assert_frame_equal, get_test_data
 
 
 TEST_DF = DataFrame(data=[
