+++ tests/rules/io/test_files_parquet.py
@@ -1,12 +1,11 @@
 import datetime
 import os
 from pandas import DataFrame
-from pandas.testing import assert_frame_equal
 import pytest
 
 from etlrules.exceptions import MissingColumnError
 from etlrules.backends.pandas import ReadParquetFileRule, WriteParquetFileRule
-from tests.backends.pandas.utils.data import get_test_data
+from tests.utils.data import assert_frame_equal, get_test_data
 
 
 TEST_DF = DataFrame(data=[
