+++ tests/rules/test_aggregate.py
@@ -1,10 +1,10 @@
 import numpy as np
 from pandas import DataFrame
-from pandas.testing import assert_frame_equal
 import pytest
 
 from etlrules.backends.pandas import AggregateRule
-from tests.backends.pandas.utils.data import get_test_data
+
+from tests.utils.data import assert_frame_equal, get_test_data
 
 
 INPUT_DF = DataFrame(data=[
