+++ tests/rules/test_basic.py
@@ -1,9 +1,8 @@
 from pandas import DataFrame
-from pandas.testing import assert_frame_equal
 import pytest
 from etlrules.backends.pandas import DedupeRule, ProjectRule, RenameRule, ReplaceRule
 from etlrules.exceptions import ColumnAlreadyExistsError, MissingColumnError
-from tests.backends.pandas.utils.data import get_test_data
+from tests.utils.data import assert_frame_equal, get_test_data
 
 
 @pytest.mark.parametrize("columns,exclude,main_input,named_inputs,named_input,named_output,expected", [
