+++ tests/rules/test_conditions.py
@@ -1,9 +1,8 @@
 from pandas import DataFrame
-from pandas.testing import assert_frame_equal
 import pytest
 from etlrules.backends.pandas import IfThenElseRule, FilterRule
 from etlrules.exceptions import ColumnAlreadyExistsError, ExpressionSyntaxError, MissingColumnError
-from tests.backends.pandas.utils.data import get_test_data
+from tests.utils.data import assert_frame_equal, get_test_data
 
 
 INPUT_DF = DataFrame(data=[
