+++ tests/rules/test_numeric.py
@@ -1,10 +1,9 @@
 from pandas import DataFrame
-from pandas.testing import assert_frame_equal
 import pytest
 
 from etlrules.exceptions import MissingColumnError
 from etlrules.backends.pandas import AbsRule, RoundRule
-from tests.backends.pandas.utils.data import get_test_data
+from tests.utils.data import assert_frame_equal, get_test_data
 
 INPUT_DF = DataFrame(data=[
     {"A": 1.456, "B": 1.456, "C": 3.8734},
