+++ tests/rules/test_serialization.py
