+++ tests/rules/test_sort.py
@@ -4,7 +4,7 @@ import pytest
 
 from etlrules.exceptions import MissingColumnError
 from etlrules.backends.pandas import SortRule
-from tests.backends.pandas.utils.data import get_test_data
+from tests.utils.data import assert_frame_equal, get_test_data
 
 
 def test_sort_rule_single_column():
