+++ tests/rules/test_strings.py
@@ -1,5 +1,4 @@
 from pandas import DataFrame
-from pandas.testing import assert_frame_equal
 import pytest
 
 from etlrules.exceptions import ColumnAlreadyExistsError, MissingColumnError
@@ -7,7 +6,7 @@ from etlrules.backends.pandas import (
     StrLowerRule, StrUpperRule, StrCapitalizeRule, StrStripRule, StrPadRule,
     StrSplitRule, StrSplitRejoinRule, StrExtractRule
 )
-from tests.backends.pandas.utils.data import get_test_data
+from tests.utils.data import assert_frame_equal, get_test_data
 
 
 INPUT_DF = DataFrame(data=[
